refactor : lectureStore를 lectureSlice로 변경
